requirements.txt
asgiref==3.11.1
black==26.1.0
click==8.3.1
colorama==0.4.6
Django==5.0.14
django-cors-headers==4.3.1
django-extensions==3.2.3
django-silk==5.4.3
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.1
gprof2dot==2025.4.14
mypy_extensions==1.1.0
packaging==26.0
pathspec==1.0.4
pillow==12.1.1
platformdirs==4.5.1
PyJWT==2.11.0
pytokens==0.4.1
sqlparse==0.5.5
tzdata==2025.3
